[gha] allow external contributors build/lint (#991)
This PR introduces `pull_request_target` for the main CI build/lint job. It requires  permissions since there are some internal packages that are used in part of the build. The `permission-check` job ensures the permissions are locked down, and Semgrep enforces that job's existence. 

This PR is tested by #992 , which should be safe to land alongside it as long as Semgrep passes on it